updates & readme

+++ database.ts
@@ -5,14 +5,14 @@ const DATABASE: string | undefined = process.env.DATABASE;
 if (!DATABASE) throw new Error("Database doesn't exists.");
 
 interface IPost {
-    date: string;
+    date: Date;
     content: string;
-    username?: string;
+    username: string;
 }
 
 const postSchema = new mongoose.Schema<IPost>({
     date: {
-        type: String,
+        type: Date,
         required: true
     },
     content: {
@@ -21,13 +21,12 @@ const postSchema = new mongoose.Schema<IPost>({
     },
     username: {
         type: String,
-        required: false,
+        required: true,
     }
 }, { collection: "posts" }); const Post = mongoose.model("Post", postSchema);
 
 export async function connectToDatabase(): Promise<void> {
-    if (!DATABASE) throw new Error("Database doesn't exists.");
-    await mongoose.connect(DATABASE, { serverSelectionTimeoutMS: 10000 });
+    await mongoose.connect(DATABASE!, { serverSelectionTimeoutMS: 10000 });
     console.log("Connected...");
 }
 
@@ -35,32 +34,17 @@ export async function loadPosts(): Promise<IPost[]> {
     return await Post.find();
 }
 
-export async function createPost(content: string, username?: string): Promise<IPost | null> {
+export async function createPost(content: string, username?: string): Promise<IPost> {
     const LIMIT = 100;
     const count = await Post.countDocuments();
 
-    if (count >= LIMIT) {
-        await Post.findOneAndDelete({}, { sort: { date: 1 } });
-    }
+    if (count >= LIMIT) await Post.findOneAndDelete({}, { sort: { date: 1 } });
 
-    const now = new Date();
-    const dateStr = now.toLocaleString("en", {
-        weekday: "long",
-        year: "numeric",
-        month: "long",
-        day: "numeric",
-        hour: "2-digit",
-        minute: "2-digit",
-        second: "2-digit",
-    });
-
-    if (!username) username = "Anonymous";
-
-    await new Post({
-        date: dateStr,
+    const newPost = await new Post({
+        date: new Date(),
         content: content,
-        username: username,
+        username: username || "Anonymous",
     }).save();
 
-    return { date: dateStr, content: content, username: username }
+    return newPost.toObject();
 }
